+++ examples/data-sources/anxcloud_availability_zone/data-source.tf
@@ -1,0 +1,4 @@
+data "anxcloud_availability_zone" "zoneA" {
+  location_id = data.anxcloud_core_location.anx04.id
+  name= "Zone A"
+}